Create LeetCode_3000. Maximum Area of Longest Diagonal Rectangle
3000. Maximum Area of Longest Diagonal Rectangle

ou are given a 2D 0-indexed integer array dimensions.

For all indices i, 0 <= i < dimensions.length, dimensions[i][0] represents the length and dimensions[i][1] represents the width of the rectangle i.

Return the area of the rectangle having the longest diagonal. If there are multiple rectangles with the longest diagonal, return the area of the rectangle having the maximum area.